started to implement Table

+++ src/table/datatype.rs
@@ -6,7 +6,7 @@ pub enum DataType {
     TinyInt {x : i8},
     Decimal {x: f32},
     Float {x: f64},
-    VarChar {x: String, y: u16},
+    VarChar {x: String, y: u8},
     Bool {x: bool},
     Date {x: i64},
     Time {x: i64},
